[WAL-110] Moved navigation logic from activity to fragment to avoid extra subjects. Reverted change made on a previous ticket
