+++ app/src/main/java/com/asfoundation/wallet/promotions/PromotionsActivity.kt
@@ -3,26 +3,15 @@ package com.asfoundation.wallet.promotions
 import android.os.Bundle
 import android.view.MenuItem
 import com.asf.wallet.R
-import com.asfoundation.wallet.router.TransactionsRouter
 import com.asfoundation.wallet.ui.BaseActivity
-import io.reactivex.subjects.PublishSubject
 
 class PromotionsActivity : BaseActivity(), PromotionsActivityView {
-
-  private lateinit var transactionsRouter: TransactionsRouter
-
-  private var backEnabled = true
-
-  private var onBackPressedSubject: PublishSubject<Any>? = null
-
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
 
     setContentView(R.layout.activity_promotions)
     toolbar()
-    onBackPressedSubject = PublishSubject.create()
-    transactionsRouter = TransactionsRouter()
     if (savedInstanceState == null) {
       supportFragmentManager.beginTransaction()
           .add(R.id.fragment_container, PromotionsFragment.newInstance())
@@ -31,25 +20,7 @@ class PromotionsActivity : BaseActivity(), PromotionsActivityView {
   }
 
   override fun onOptionsItemSelected(item: MenuItem): Boolean {
-    return if (item.itemId == android.R.id.home) {
-      if (backEnabled) {
-        transactionsRouter.open(this, true)
-      } else {
-        onBackPressedSubject?.onNext("")
-      }
-      true
-    } else {
-      super.onOptionsItemSelected(item)
-    }
-  }
-
-  override fun backPressed() = onBackPressedSubject!!
-
-  override fun enableBack() {
-    backEnabled = true
-  }
-
-  override fun disableBack() {
-    backEnabled = false
+    //Let Fragment handle the click
+    return false
   }
 }
