+++ app/src/main/java/com/asfoundation/wallet/promotions/PromotionsActivityView.kt
@@ -1,12 +1,3 @@
 package com.asfoundation.wallet.promotions
 
-import io.reactivex.Observable
-
-interface PromotionsActivityView {
-
-  fun backPressed(): Observable<Any>
-
-  fun enableBack()
-
-  fun disableBack()
-}
+interface PromotionsActivityView
