+++ app/src/main/java/com/asfoundation/wallet/ui/backup/BackupActivity.kt
@@ -3,9 +3,11 @@ package com.asfoundation.wallet.ui.backup
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
+import android.view.MenuItem
 import com.asf.wallet.R
 import com.asfoundation.wallet.permissions.manage.view.ToolbarManager
 import com.asfoundation.wallet.ui.BaseActivity
+import com.asfoundation.wallet.ui.backup.entry.BackupWalletFragment
 import dagger.android.AndroidInjection
 import io.reactivex.subjects.PublishSubject
 import javax.inject.Inject
@@ -35,6 +37,13 @@ class BackupActivity : BaseActivity(), BackupActivityView, ToolbarManager {
     presenter.present(savedInstanceState == null)
   }
 
+  override fun onOptionsItemSelected(item: MenuItem): Boolean {
+    if (item.itemId == android.R.id.home) {
+      presenter.sendWalletSaveFileEvent(getCurrentFragment())
+    }
+    return super.onOptionsItemSelected(item)
+  }
+
   override fun closeScreen() = finish()
 
   override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
@@ -50,5 +59,11 @@ class BackupActivity : BaseActivity(), BackupActivityView, ToolbarManager {
 
   override fun setupToolbar() {
     toolbar()
+  }
+
+  private fun getCurrentFragment(): String {
+    val fragments = supportFragmentManager.fragments
+    return if (fragments.isNotEmpty()) fragments[0]::class.java.simpleName
+    else BackupWalletFragment::class.java.simpleName
   }
 }
