+++ app/src/main/java/com/asfoundation/wallet/ui/backup/creation/BackupCreationPresenter.kt
@@ -50,11 +50,6 @@ class BackupCreationPresenter(private val view: BackupCreationView,
     handleSystemFileIntentResult()
   }
 
-  fun sendWalletSaveFileCanceledEvent() {
-    walletsEventSender.sendWalletSaveFileEvent(WalletsAnalytics.ACTION_BACK,
-        WalletsAnalytics.STATUS_FAIL, WalletsAnalytics.REASON_CANCELED)
-  }
-
   private fun handleSystemFileIntentResult() {
     disposables.add(view.onSystemFileIntentResult()
         .observeOn(networkScheduler)
